remove gnome-print include
2003-07-02  Jody Goldberg <jody@gnome.org>

	* widget-font-selector.c : remove gnome-print include

	* widget-pixmap-combo.h : Cleanup the includes

	* gtk-combo-box.c (gtk_combo_box_init) : Use GObject properties in
	  place of deprecated gtk_window_set_policy.
	(gtk_combo_popup_tear_off) : ditto.
	(gtk_combo_box_popup_display) : Use gtk_window_move in place of
	  deprecated gtk_window_set_uposition.
	(gtk_combo_box_get_pos) : Make multi head safe.

2003-07-02  Jody Goldberg <jody@gnome.org>

	* dialog-tabulate.c (dialog_tabulate) : Use GObject properties in
	  place of deprecated gtk_window_set_policy.

2003-07-02  Jody Goldberg <jody@gnome.org>

	* src/sheet-control-gui.c (scg_comment_display) : Use gtk_window_move
	  in place of the deprecated gtk_widget_set_uposition.

+++ src/dialogs/dialog-stf.c
@@ -347,7 +347,7 @@ stf_dialog (WorkbookControlGUI *wbcg,
 	pagedata.wbcg = wbcg;
 	pagedata.source = source;
 	pagedata.raw_data = data;
-	pagedata.raw_data_len = data_len < 0 ? strlen (data) : data_len;
+	pagedata.raw_data_len = data_len < 0 ? (int)strlen (data) : data_len;
 	pagedata.utf8_data = NULL;
 	pagedata.cur = NULL;
